Bot/tg_bot/requirements.txt
blinker==1.8.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
CurrencyConverter==0.17.25
Faker==25.3.0
Flask==3.0.3
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
pyTelegramBotAPI==4.9.0
python-dateutil==2.9.0.post0
python-dotenv==0.21.1
requests==2.31.0
six==1.16.0
urllib3==2.2.1
Werkzeug==3.0.3
